bug fixes, and throttler

+++ src/flats/flats.controller.ts
@@ -26,7 +26,9 @@ import {AddGPTAnswersDto} from "./dto/add-gpt-answers.dto";
 import {FlatsGPT} from "./entities/flats-gpt.entity";
 import {DeleteResult} from "typeorm";
 import {SkipThrottle} from "@nestjs/throttler";
-@SkipThrottle({ long: true })
+
+// long:true means use short!
+@SkipThrottle({ short: true })
 @Controller('/api/flats')
 export class FlatsController {
